# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkillUp LMS Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new freelancer or client account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /freelancer/profile:
    get:
      operationId: getFreelancerProfile
      tags: [freelancer]
      summary: Get my freelancer profile
      responses:
        "200":
          description: Freelancer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerProfile"
    put:
      operationId: upsertFreelancerProfile
      tags: [freelancer]
      summary: Create or update my freelancer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreelancerProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerProfile"
  /freelancer/bank-details:
    get:
      operationId: getBankDetails
      tags: [freelancer]
      summary: Get my bank details
      responses:
        "200":
          description: Bank details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankDetails"
    put:
      operationId: upsertBankDetails
      tags: [freelancer]
      summary: Create or update my bank details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankDetailsInput"
      responses:
        "200":
          description: Updated bank details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankDetails"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [freelancer, client]
        phone:
          type: string
    AuthUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, trainer, student, freelancer, client]
...
    FreelancerProfile:
      type: object
      required: [id, userId]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        hourlyRate:
          type: ["number", "null"]
        location:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        availability:
          type: string
          enum: [available, busy, unavailable]
    FreelancerProfileInput:
      type: object
      properties:
        title:
          type: string
        bio:
          type: string
        skills:
          type: string
        hourlyRate:
          type: number
        location:
          type: string
        portfolioUrl:
          type: string
        availability:
          type: string
          enum: [available, busy, unavailable]
...
    BankDetails:
      type: object
      required: [id, userId, accountHolder, accountNumber, ifscCode, bankName]
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountHolder:
          type: string
        accountNumber:
          type: string
        ifscCode:
          type: string
        bankName:
          type: string
        branch:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
    BankDetailsInput:
      type: object
      required: [accountHolder, accountNumber, ifscCode, bankName]
      properties:
        accountHolder:
          type: string
        accountNumber:
          type: string
        ifscCode:
          type: string
        bankName:
          type: string
        branch:
          type: string
        upiId:
          type: string
...