# lib/api-zod/src/generated/types/bankDetails.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillUp LMS Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface BankDetails {
  id: number;
  userId: number;
  accountHolder: string;
  accountNumber: string;
  ifscCode: string;
  bankName: string;
  /** @nullable */
  branch?: string | null;
  /** @nullable */
  upiId?: string | null;
}