# lib/api-zod/src/generated/types/bankDetailsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillUp LMS Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface BankDetailsInput {
  accountHolder: string;
  accountNumber: string;
  ifscCode: string;
  bankName: string;
  branch?: string;
  upiId?: string;
}