# lib/api-zod/src/generated/types/freelancerProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillUp LMS Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { FreelancerProfileInputAvailability } from './freelancerProfileInputAvailability';

export interface FreelancerProfileInput {
  title?: string;
  bio?: string;
  skills?: string;
  hourlyRate?: number;
  location?: string;
  portfolioUrl?: string;
  availability?: FreelancerProfileInputAvailability;
}